Modularize cv functionalities
